+++ src/Runner/Proof/Properties.php
@@ -84,7 +84,6 @@ final class Properties implements Proof
             $this->systemUnderTest,
         )
             ->immutable()
-            ->toSet()
             ->randomize()
             ->take($count);
     }